Prune two examples and tune sky-blur defaults
Remove the Cuboid and Instancing examples (SpinComponent moves to the
render-target example, its only other user). Start the Stress Tests
scene at 0.40 sky blur on the Helipad environment, and the Car at 0.60
sky blur.
