--- examples/flutter_app/lib/example_cuboid.dart
@@ -1,65 +1,0 @@
-import 'dart:math';
-
-import 'package:flutter/material.dart';
-import 'package:flutter_scene/scene.dart';
-import 'package:vector_math/vector_math.dart' as vm;
-
-import 'example_settings.dart';
-
-class ExampleCuboid extends StatefulWidget {
-  const ExampleCuboid({super.key});
-
-  @override
-  ExampleCuboidState createState() => ExampleCuboidState();
-}
-
-/// Spins the owning node about its Y axis. A minimal user-defined
-/// [Component]: behavior attached to a node through the [update] hook.
-class SpinComponent extends Component {
-  SpinComponent(this.radiansPerSecond);
-
-  final double radiansPerSecond;
-
-  @override
-  void update(double deltaSeconds) {
-    node.localTransform =
-        node.localTransform *
-        vm.Matrix4.rotationY(radiansPerSecond * deltaSeconds);
-  }
-}
-
-class ExampleCuboidState extends State<ExampleCuboid> {
-  Scene scene = Scene();
-
-  @override
-  void initState() {
-    final mesh = Mesh(
-      CuboidGeometry(vm.Vector3(1, 1, 1), debugColors: true),
-      UnlitMaterial(),
-    );
-    // The camera orbits at 1 rad/s, so the spin rate is chosen to differ
-    // from it; otherwise the cube would appear stationary relative to the
-    // camera.
-    final node = Node(mesh: mesh)..addComponent(SpinComponent(-1.5));
-    scene.add(node);
-
-    super.initState();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    // SceneView owns the frame ticker and the render call; the spin component
-    // advances via Scene.render's implicit per-frame tick.
-    return SceneView(
-      scene,
-      cameraBuilder: (elapsed) {
-        final t = elapsed.inMicroseconds / 1e6;
-        return PerspectiveCamera(
-          position: vm.Vector3(sin(t) * 5, 2, cos(t) * 5),
-          target: vm.Vector3(0, 0, 0),
-        );
-      },
-      onTick: (elapsed, deltaSeconds) => exampleSettings.applyTo(scene),
-    );
-  }
-}
